General application assets bootstrapping.
