+++ django/ai/static/css/admin.css
