+++ django/ai/static/css/base.css
