+++ django/ai/templates/base.html
@@ -8,17 +8,64 @@
 
 		<title>{% block subtitle %}{% endblock %}{{ site.name }}</title>
 
-		<link rel="stylesheet" href="{% static "base.css" %}" type='text/css'>
-		<link rel="stylesheet" href="{% static "admin.css" %}" type='text/css'>
+		<link rel="stylesheet" href="{% static 'css/base.css' %}" type='text/css'>
+		<link rel="stylesheet" href="{% static 'css/admin.css' %}" type='text/css'>
+		<link rel="stylesheet" href="{% static 'vendor/css/semantic.min.css' %}" type='text/css'>
 
 		{% block head-extras-app %}{% endblock %}
 		{% block head-extras-page %}{% endblock %}
 	</head>
 	<body class="{% block body-classes %}{% endblock %}">
 		<div id="page-component"></div>
-		<script src="{% static 'potassium.js' %}"></script>
-		<script src="{% static 'base.js' %}"></script>
-		<div id="content">{% block content%}{% endblock %}</div>
+
+    <script crossorigin src="https://unpkg.com/react@16/umd/react.production.min.js"></script>
+    <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js"></script>
+
+    <script
+			src="http://code.jquery.com/jquery-2.2.4.min.js"
+			integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44="
+			crossorigin="anonymous"></script>
+
+		<script src="{% static 'vendor/js/semantic.min.js' %}"></script>
+		<script src="{% static 'js/base.js' %}"></script>
+
+		<div id="content">
+
+      <div class="ui menu">
+        <div class="header item">Art::Infrastructure</div>
+        <a class="item" href="/">Dashboard</a>
+        <a class="item" href="/">Artwork</a>
+
+        <div class="ui dropdown item" tabindex="0">
+          Lighting
+          <i class="dropdown icon"></i>
+          <div class="menu transition hidden" tabindex="-1">
+            <div class="item" href='{% url "lighting:bacnet_lights" %}'>BACnet light</div>
+            <div class="divider"></div>
+            <div class="item" href='{% url "lighting:projectors" %}'>Projectors</div>
+            <div class="divider"></div>
+            <div class="item" href='{% url "lighting:crestons" %}'>Creston</div>
+          </div>
+        </div>
+
+        <a class="item" href="/">Scheduler</a>
+        <a class="item" href="/">Weather</a>
+        <a class="item" href='{% url "heartbeat:index" %}'>System Status</a>
+        <div class="right menu">
+          <div class="item">
+            <div class="ui action left icon input">
+              <input placeholder="Search" type="text">
+              <button class="ui button">Search</button>
+            </div>
+          </div>
+        </div>
+        <a class="item" href="/admin/">Admin</a>
+      </div>
+
+      {% block content%}{% endblock %}
+
+    </div>
+
 		{% block script-extras-app %}{% endblock %}
 		{% block script-extras-page %}{% endblock %}
 	</body>
