--- django/front/static/front/front.css
@@ -1,4 +1,0 @@
-
-.front-index-page-component .links-row { margin-top: 30px; }
-.front-index-page-component .links-row h2 { margin-bottom: 10px; color: blue; }
-.front-index-page-component .links-row p { margin-bottom: 30px; }
