+++ django/front/templates/front/base.html
@@ -2,9 +2,7 @@
 {% load staticfiles %}
 
 {% block head-extras-app %}
-	<link rel="stylesheet" href="{% static "front/front.css" %}" type='text/css'>
 {% endblock %}
 
 {% block script-extras-app %}
-	<script src="{% static 'front/front.js' %}"></script>
 {% endblock %}
