--- django/heartbeat/static/heartbeat/heartbeat.js
@@ -1,13 +1,0 @@
-var heartbeat = heartbeat || {}
-heartbeat.ui = heartbeat.ui || {}
-
-heartbeat.ui = heartbeat.ui || {}
-
-heartbeat.ui.IndexPageComponent = class extends k.Component {
-	constructor(dataObject=null, options={}){
-		super(dataObject, options)
-
-		this.topNav = new base.ui.TopNavComponent()
-		this.el.appendChild(this.topNav.el)
-	}
-}
