+++ django/heartbeat/templates/heartbeat/base.html
@@ -2,9 +2,8 @@
 {% load staticfiles %}
 
 {% block head-extras-app %}
-	<link rel="stylesheet" href="{% static "heartbeat/heartbeat.css" %}" type='text/css'>
 {% endblock %}
 
 {% block script-extras-app %}
-	<script src="{% static 'heartbeat/heartbeat.js' %}"></script>
+    <script src="{% static 'js/heartbeat.js' %}"></script>
 {% endblock %}
