+++ django/lighting/templates/lighting/.#bacnet_lights.html
@@ -1,0 +1,1 @@
+devy@localhost.localdomain.12645:1524525708
