+++ django/lighting/templates/lighting/base.html
@@ -2,9 +2,7 @@
 {% load staticfiles %}
 
 {% block head-extras-app %}
-	<link rel="stylesheet" href="{% static "lighting/front.css" %}" type='text/css'>
 {% endblock %}
 
 {% block script-extras-app %}
-	<script src="{% static 'lighting/lighting.js' %}"></script>
 {% endblock %}
