+++ django/lighting/templates/lighting/projectors.html
@@ -1,0 +1,44 @@
+{% extends "base.html" %}
+
+{% block subtitle %} {% endblock %}
+
+{% block style %}{% endblock %}
+
+{% block sub-head %}{% endblock %}
+
+{% block content%}
+
+<div class="ui container">
+
+  <h2 class="ui dividing header">Lighting::Projectors</h2>
+
+  <div class="ui one column relaxed grid">
+    <div class="column">
+
+      <table class="ui very basic right aligned table">
+        <thead>
+          <tr>
+            <th class="left aligned">Projector</th>
+            <th>Actions</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td class="left aligned">Projector #1</td>
+            <td>view</td>
+          </tr>
+          <tr>
+            <td class="left aligned">Projector #1</td>
+            <td>view</td>
+          </tr>
+          <tr>
+            <td class="left aligned">Projector #1</td>
+            <td>view</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+
+</div>
+{% endblock %}
